+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 0.8.0
+- Make use of `babel-runtime`
+- If Babel is available, parse Riot tags with it
+- **BREAKING** Drop runtime check for Babel
+  - Use `.babelrc` instead
+
 ### 0.7.0
 - Use Babel fully for ES2015+
 - Replace JSHint with ESLint